@@ -19,31 +19,35 @@
   "author": "Cristina Shaver (https://github.com/cshaver)",
   "license": "MIT",
   "dependencies": {
+    "core-js": "^2.5.3",
     "firebase": "^4.5.2",
     "firebaseui": "^2.4.1",
     "lodash": "^4.17.4",
     "moment": "^2.19.1",
     "normalize.css": "^7.0.0",
     "object-assign": "^4.1.1",
-    "promise": "^7.1.1",
+    "promise": "^8.0.1",
     "prop-types": "^15.5.10",
-    "react": "^15.5.4",
-    "react-dom": "^15.5.4",
+    "raf": "^3.4.0",
+    "react": "^16.2.0",
+    "react-dom": "^16.2.0",
     "react-redux": "^5.0.4",
     "react-redux-firebase": "next",
-    "react-router": "^3.0.0",
+    "react-router": "^4.2.0",
+    "react-router-dom": "^4.2.2",
     "recompose": "^0.26.0",
     "redbox-react": "^1.3.6",
     "reduce-reducers": "^0.1.2",
     "redux": "^3.6.0",
-    "redux-auth-wrapper": "^1.0.0",
-    "redux-form": "^6.6.1",
+    "redux-auth-wrapper": "^2.0.2",
+    "redux-form": "^7.2.0",
     "redux-thunk": "^2.2.0",
     "slugify": "^1.2.1",
     "whatwg-fetch": "^2.0.3"
   },
   "devDependencies": {
-    "babel-core": "^6.26.0",
+    "@babel/core": "^7.0.0-beta.35",
+    "@babel/plugin-transform-react-jsx": "^7.0.0-beta.35",
     "babel-loader": "^7.1.2",
     "babel-plugin-transform-decorators-legacy": "^1.3.4",
     "babel-preset-airbnb": "^2.4.0",
