@@ -3,7 +3,7 @@ import React from 'react';
 export const NotFound = () => (
   <div>
     <h1>Whoops! 404!</h1>
-    <p>This page was not found.</p>
+    <p>You must be lost. This page was not found.</p>
   </div>
 );
 
