@@ -1,5 +1,5 @@
 import React from 'react';
-import { Link } from 'react-router';
+import { Link } from 'react-router-dom';
 
 const ProfileLink = ({ profile, children }) => (
   <Link to={`/profile/${profile.uid}/${profile.slug}`}>
