@@ -1,7 +1,7 @@
+import { withRouter } from 'react-router';
+
 import Feed from '../../containers/Feed';
 import { UserIsAuthenticated } from '../../utils/router';
 
 // Sync route definition
-export default {
-  component: UserIsAuthenticated(Feed),
-};
+export default withRouter(UserIsAuthenticated(Feed));
